+++ src/mqtt.h
@@ -163,7 +163,7 @@ class Mqtt {
         return mqtt_message_id_;
     }
 
-    static uint32_t publish_fails() {
+    static uint16_t publish_fails() {
         return mqtt_publish_fails_;
     }
